Implemented schedule table output logic. Made changes to repositories and formatters according to the new id type. Lesson number and day of week implemented as enums.

+++ src/main/java/scheduleApp/AuditoriumRepository.java
@@ -2,6 +2,6 @@ package scheduleApp;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuditoriumRepository extends JpaRepository<Auditorium, Long> {
+public interface AuditoriumRepository extends JpaRepository<Auditorium, Integer> {
 
 }
