+++ src/main/java/scheduleApp/InstructorRepository.java
@@ -2,6 +2,6 @@ package scheduleApp;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InstructorRepository extends JpaRepository<Instructor, Long> {
+public interface InstructorRepository extends JpaRepository<Instructor, Integer> {
 
 }
